(api-break) rename 'utils' -> '_utils' and it's automatically imported in __init__.py
