+++ examples/popping_widget_1.py
@@ -4,7 +4,7 @@ from kivy.app import App
 from kivy.lang import Builder
 from kivy.graphics import Rotate, Translate
 from asynckivy import vanim
-from asynckivy.utils import transform
+from asynckivy import transform
 
 
 @contextmanager
