+++ examples/popping_widget_3_elastic.py
@@ -7,7 +7,7 @@ from kivy.app import App
 from kivy.lang import Builder
 from kivy.graphics import Translate, Scale
 from asynckivy import vanim
-from asynckivy.utils import transform
+from asynckivy import transform
 
 
 @contextmanager
