+++ src/asynckivy/__init__.py
@@ -1,5 +1,5 @@
 from asyncgui import *
-from .exceptions import *
+from ._exceptions import *
 from ._sleep import *
 from ._event import *
 from ._animation import *
@@ -7,3 +7,4 @@ from ._interpolate import *
 from ._touch import *
 from ._threading import *
 from ._n_frames import *
+from ._utils import *
