+++ src/asynckivy/_exceptions.py
@@ -1,0 +1,7 @@
+__all__ = (
+    'MotionEventAlreadyEndedError',
+)
+
+
+class MotionEventAlreadyEndedError(Exception):
+    """A MotionEvent has already ended."""
