--- src/asynckivy/exceptions.py
@@ -1,26 +1,0 @@
-__all__ = (
-    'MotionEventAlreadyEndedError',
-    'WouldBlock', 'ClosedResourceError', 'EndOfResource',
-)
-
-
-class MotionEventAlreadyEndedError(Exception):
-    """A MotionEvent has already ended."""
-
-
-class WouldBlock(Exception):
-    """(took from trio)
-    Raised by X_nowait functions if X would block.
-    """
-
-
-class ClosedResourceError(Exception):
-    """(took from trio)
-    Raised when attempting to use a resource after it has been closed.
-    """
-
-
-class EndOfResource(Exception):
-    """(similar to trio's EndOfChannel)
-    This is analogous to an "end-of-file" condition, but for resources in general.
-    """
